Iteration #1
This first iteration set’s up the first component of my device. No issues here as it was all relatively straightforward. All this circuit does is Display the temperature and humidity sensor on an LCD. The LCD I have used before, but the DHT-11 was new to me so I learnt about how to use it in the elegoo documentation.