Iteration #2
This second iteration, much like the previous one, sets up for my future plans. I have added an RTC (Real Time Clock) that once code has been uploaded to it, will hold time. The date and time are displayed on the LCD and the Serial Monitor.
One slight Issue I have encountered with this module is that if unplugged, the battery in the RTC doesn’t keep the module on and running. I have started researching the reason behind this and it turns out that the command that I use for the RTC will only operate when power is running through both sides of the RTC circuit. I will experiment with some ways to fix this issue, but in a final product, the circuit would never be disconnected from a constant source of power so it wouldn’t be an issue.