Iteration #3
Iterations #3 and #3.1 were experimental iterations where I aimed to use the shift register to control a 4-digit 7-segment display. This display would have shown the time directly from the RTC. Unfortunately, both attempts at integrating the display were unsuccessful. The reason why I aimed to use both an LCD and the 4x7 display was that it would have demonstrated how a shift register can be used to allow multiple devices that require lots of pins to be used simultaneously.

Iteration #3 was using the code example from the elegoo documentation to try and display the RTC data as a hex code digit from the character table. This did not turn out well, so the code for this Iteration I have provided has the 4x7 display connected, but will not display anything for the reason that it would break.