Iteration #3.1
Iterations #3 and #3.1 were experimental iterations where I aimed to use the shift register to control a 4-digit 7-segment display. This display would have shown the time directly from the RTC. Unfortunately, both attempts at integrating the display were unsuccessful. The reason why I aimed to use both an LCD and the 4x7 display was that it would have demonstrated how a shift register can be used to allow multiple devices that require lots of pins to be used simultaneously.

Iteration #3.1 is a slightly different attempt to Iteration #3 as I tried to use the code from this webpage: https://github.com/rmorenojr/ElegooTutorial/tree/master/Lesson%2028%20-%204%20Digit%207%20Segment%20Display
In an attempt to fix my issue. Unfortunately, due to issues with the separate delay structures for my original circuit and the newly-integrated code, I could only use one device or the other (those devices being the LCD or the 4x7 display). This meant that I had to give up on the idea of using the 4x7 Display as I couldn’t find any way to use both devices without needing multiple shift registers to control everything with one large loop.
The code Displayed here will only work if either the 4x7’s code is noted out or vice-versa for the LCD’s code as only one will be able to operate at a time.
One funky thing I did borrow from this person's code was the Titles for the code to better lay it out.