Iteration #4
This iteration is a much better alternative to the failed 3.0 iterations in my opinion. I have taken ideas from the last iteration to simplify my product into one smaller device. This way it will more power efficient and easier to handle. 
What I have done is used the shift register to control the LCD, meaning the LCD only takes up 3 pins on the Arduino. This is an effective use of the shift register because it demonstrates how it can be used in a real-world application to control large devices that take up lots of pins.
The DHT-11 Temperature & Humidity sensor as well as the DS1307 RTC Module gather data and send it to the Arduino. The Arduino then sends that data to the shift register where it is stored and sent out to the LCD. Because the time and temperature are constantly changing variables, new data will repeatedly be sent to the shift register to send to the LCD. 

Usually, when connecting an LCD to the Arduino the connections have to be physical, which is not what is going on in this circuit as a shift register is between the LCD and the Arduino. To solve this issue, a modified version of the Liquid Crystal Library has to be used. This library has the exact same functions as the original one but allows for a shift register to be used to save up I/O pins on the microcontroller. 
The modified library can be found here: https://playground.arduino.cc/Main/LiquidCrystal/
And I used this guide to help me set up the pinout of the LCD correctly: https://www.instructables.com/LCD-Screen-With-74HC595-and-Keypad-Module/